requirements.txt
playwright==1.41.0
python-dotenv==1.0.1
pandas==2.2.0
openpyxl==3.1.2
beautifulsoup4==4.12.3
fake-useragent==1.5.1
loguru==0.7.2
